+++ kernel/heap_setup.c
@@ -187,7 +187,7 @@ static void vdefrag(struct heap_list *heap_entry)
 	if (heap_entry->page_size == start->size + sizeof(struct alloc_header) && start->free)
 	{
 		list_del(&heap_entry->list);
-		for (void *tmp_virt_addr = heap_entry ; tmp_virt_addr < heap_entry->page_size ; tmp_virt_addr += 4096)
+		for (void *tmp_virt_addr = heap_entry ; (size_t)tmp_virt_addr < heap_entry->page_size ; tmp_virt_addr += 4096)
 			free_phys_block(page_get_phys(tmp_virt_addr), 1);
 		free_virt_block(heap_entry, heap_entry->page_size);
 	}
